move ClawSubsystem to intake subfolder and rename
