+++ src/main/java/frc/robot/commands/CubeGrabState.java
@@ -2,13 +2,13 @@ package frc.robot.commands;
 
 import java.util.function.DoubleSupplier;
 
-import frc.robot.subsystems.ClawSubsystem;
+import frc.robot.subsystems.intake.IntakeSubsytstem;
 import frc.team4272.globals.State;
 
-public class CubeGrabState extends State<ClawSubsystem> {
+public class CubeGrabState extends State<IntakeSubsytstem> {
     private DoubleSupplier power;
 
-    public CubeGrabState(ClawSubsystem claw, DoubleSupplier power) {
+    public CubeGrabState(IntakeSubsytstem claw, DoubleSupplier power) {
         super(claw);
         this.power = power;
     }
