+++ src/main/java/frc/robot/subsystems/intake/IntakeSubsytstem.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems;
+package frc.robot.subsystems.intake;
 
 import com.revrobotics.CANSparkMax;
 import com.revrobotics.CANSparkMaxLowLevel.MotorType;
@@ -11,12 +11,12 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
 import static frc.robot.constants.HardwareMap.*;
 
-public class ClawSubsystem extends SubsystemBase {
+public class IntakeSubsytstem extends SubsystemBase {
     private CANSparkMax clawLeaderMotor = new CANSparkMax(CLAW_RIGHT_ID, MotorType.kBrushless);
     private CANSparkMax clawFollowerMotor = new CANSparkMax(CLAW_LEFT_ID, MotorType.kBrushless);
 
     /** Creates a new ClawSubsystem. */
-    public ClawSubsystem() {
+    public IntakeSubsytstem() {
         clawLeaderMotor.restoreFactoryDefaults();
         clawFollowerMotor.restoreFactoryDefaults();
